+++ CHANGELOG.md
@@ -20,6 +20,7 @@
   * Added a feature to render datatype.
   * Implemented sending events by topics.
   * Added ability to set a theme via URL `setTheme` search parameter.
+  * Every app now uses dynamic routes.
 
 ## v0.1.7 (2022-06-28)
   * Created database if it does not exist before migraions.